Rename esmvalcore.types to esmvalcore.typing
