+++ esmvalcore/typing.py
